requirements.txt
python-telegram-bot==21.7
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai>=0.3.0
python-dotenv==1.0.1
lxml==5.1.0
httpx>=0.24.0
pandas>=2.0.0
openpyxl>=3.1.0

